requirements.txt
numpy

PyQt4

pyserial